backend/requirements.txt
supabase-py==2.3.0
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
gunicorn==21.2.0
